requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# ML / Fine-tuning
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0
bitsandbytes>=0.43.0
accelerate>=0.28.0

# LLM API clients
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
httpx>=0.27.0

# Evaluation
scikit-learn>=1.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
